Refactor card poller
Card status poller is refactored in a separate goroutines. Singleton structure `ReaderPoller` updates list of readers, and waits for changes on currently selected reader. It also keeps state of the selected reader, and notifies `cardLoop` of reader changes.

Select widget in the toolbar doesn't now contains minimal logic for readers. It is only used for sending reader change events to the `ReaderPoller`.